+++ myserver/include/conf/vhost/vhost_manager.h
@@ -29,7 +29,7 @@ public:
   VhostManagerHandler ();
   virtual ~VhostManagerHandler ();
   virtual Vhost* getVHost (const char*, const char*, u_short);
-  virtual Vhost* getVHost (int n);
+  virtual Vhost* getVHost (u_long n);
   virtual int load (const char *resource){return 0;}
 };
