+++ myserver/include/filter/filter.h
@@ -36,11 +36,11 @@ public:
   {
     protocol = pr;
   }
-  virtual int getHeader (char* buffer, u_long len, u_long* nbw);
-  virtual int getFooter (char* buffer, u_long len, u_long* nbw);
-  virtual int read (char* buffer, u_long len, u_long*);
-  virtual int write (const char* buffer, u_long len, u_long*);
-  virtual int flush (u_long*);
+  virtual int getHeader (char* buffer, size_t len, size_t* nbw);
+  virtual int getFooter (char* buffer, size_t len, size_t* nbw);
+  virtual int read (char* buffer, size_t len, size_t *);
+  virtual int write (const char* buffer, size_t len, size_t *);
+  virtual int flush (size_t *);
   virtual int modifyData ();
   virtual const char* getName ();
   void setParent (Stream*);
