+++ myserver/include/filter/stream.h
@@ -36,9 +36,9 @@ typedef int FileHandle;
 class Stream
 {
 public:
-  virtual int read (char* buffer, u_long len, u_long*);
-  virtual int write (const char* buffer, u_long len, u_long*);
-  virtual int flush (u_long*);
+  virtual int read (char *buffer, size_t len, size_t *);
+  virtual int write (const char *buffer, size_t len, size_t *);
+  virtual int flush (size_t *);
   virtual Handle getHandle ();
   virtual int close ();
   Stream ();
